pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jimeng_visual_generation"
version = "0.1.0"
description = "MCP server for Jimeng Visual Generation API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]",
    "httpx",
    "python-dotenv"
]

[project.scripts]
jimeng_visual_generation = "jimeng_visual_generation.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/jimeng_visual_generation"]
